Define function for the creation of books
All books inside the Book construct will be passed 4 arguments.
Title, author, pages, and whether or not the book has been read.
Additional to this, all books will be given a unique id.

Also, a function for the creation of books, and their addition to the
library array has been declared. All books are pushed inside the array
after their creation.