Add function to display books in the html table
First, the table was given a class which was selected in the js file
as booksTable.
Then, a function called displayBooks was declared.
Inside this function there is a while conditional, which states that
as long as the books array has a length higher than 1, all rows after
the first one will be deleted from booksTable.

Then, a forEach loop is used to iterate over all the keys/properties of
each book element inside the library array.