Add functionality to modal dialog
An event listener to the submit event was declared. This event will
place a new book object inside the library array using the values
placed in each of the inputs from the form.

Also, after the info has been submitted, the inputs of the form are
reset and the dialog modal is closed. There's also a cancel button
for doing exactly that.

A prevent default was also added to the button.