Add toggle read button
A new method was added to the Book object prototype. Said method is to
allow changing the content of the 'read' property from the book objects
from Yes to No an viceversa.

Basically the same event listener delegation was used for the activation
of the toggle read button, just that this one invokes a new function
called toggleRead.

This function will use the find method to get the book object that has
the same id as the one that had its 'toggle read' button clicked.
Then, it invokes the toggleReadStatus prototype method to change its
read status.